# api/openapi/nsv-api-documentation.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/posts": {
      "get": {
        "operationId": "PostController_getPublishedPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "authorId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "categoryIds",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "createdAtFrom",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "createdAtTo",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "updatedAtFrom",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "updatedAtTo",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized paginated result of PostBriefResponseDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostBriefResponseDto"
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      },
      "post": {
        "operationId": "PostController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostBaseCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The serialized result of Creation result",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OperationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/all": {
      "get": {
        "operationId": "PostController_getAllPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "authorId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "categoryIds",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "createdAtFrom",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "createdAtTo",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "updatedAtFrom",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "updatedAtTo",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized paginated result of PostBriefResponseDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostBriefResponseDto"
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/{id}": {
      "get": {
        "operationId": "PostController_getPostById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized result of PostDetailResponseDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostDetailResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      },
      "put": {
        "operationId": "PostController_updatePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostBaseUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The serialized result of OperationMetadataDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OperationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      },
      "delete": {
        "operationId": "PostController_deletePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized result of OperationMetadataDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OperationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/{id}/publish": {
      "put": {
        "operationId": "PostController_publishPost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized result of OperationMetadataDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OperationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/{id}/unpublish": {
      "put": {
        "operationId": "PostController_unpublishPost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The serialized result of OperationMetadataDto",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SerializedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OperationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    }
  },
  "info": {
    "title": "NSV - NSV API Documentation API",
    "description": "The NSV API Documentation API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SerializedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "number"
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "message",
          "statusCode",
          "timestamp"
        ]
      },
      "UserBaseResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "image": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "email"
        ]
      },
...
      "PaginationMetadataDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "totalPages",
          "page",
          "limit"
        ]
      },
      "OperationMetadataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "success"
        ]
      },
...
      "PostBriefResponseDto": {
        "type": "object",
        "properties": {
          "author": {
            "$ref": "#/components/schemas/UserBaseResponseDto"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "thumbnail": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "published": {
            "type": "boolean"
          }
        },
        "required": [
          "author",
          "id",
          "title",
          "slug",
          "published"
        ]
      },
      "PostDetailResponseDto": {
        "type": "object",
        "properties": {
          "author": {
            "$ref": "#/components/schemas/UserBaseResponseDto"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "thumbnail": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "published": {
            "type": "boolean"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "author",
          "id",
          "title",
          "slug",
          "published",
          "content"
        ]
      },
      "PostBaseCreateDto": {
        "type": "object",
        "properties": {
          "authorId": {
            "type": "string"
          },
          "publishedAt": {
            "type": "number"
          },
          "published": {
            "type": "boolean"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "content": {
            "type": "string",
            "minLength": 1
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "thumbnail": {
            "type": "string",
            "format": "uri"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "title",
          "slug",
          "content"
        ]
      },
      "PostBaseUpdateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "content": {
            "type": "string",
            "minLength": 1
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "thumbnail": {
            "type": "string",
            "format": "uri"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "authorId": {
            "type": "string"
          },
          "publishedAt": {
            "type": "number"
          },
          "published": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}